# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями, комплектациями, опциями и ценами",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить список всех автомобилей",
                "operationId": "1fd4f1fdc0da258c5084bcdce4bbe693",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Toyota Camry"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Создать новый автомобиль",
                "operationId": "6f519ab2c9767d4d3b1be7cf8f1287ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toyota Corolla"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Автомобиль создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Toyota Corolla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить информацию об автомобиле",
                "operationId": "b222bfa902eefbb6a4a875762081dcbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об автомобиле",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Toyota Camry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Автомобиль не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Обновить данные автомобиля",
                "operationId": "1d232cedb5f39024dccd947cddd10087",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Toyota Camry Updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Автомобиль обновлён"
                    },
                    "404": {
                        "description": "Автомобиль не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Удалить автомобиль",
                "operationId": "802b64f09e484fbb1bb3817362d30cc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Автомобиль удалён"
                    },
                    "404": {
                        "description": "Автомобиль не найден"
                    }
                }
            }
        },
...
    },
...
}